chore(telemetry): move telemetry into its own package
