+++ telemetry/usage.go
@@ -1,34 +1,35 @@
-package prometheus
+package telemetry
 
 import (
 	"context"
 
 	platform "github.com/influxdata/influxdb"
+	pr "github.com/influxdata/influxdb/prometheus"
 	"github.com/prometheus/client_golang/prometheus"
 )
 
 var _ platform.UsageService = (*UsageService)(nil)
 
-var defaultMatcher = NewMatcher().
+var defaultMatcher = pr.NewMatcher().
 	Family("influxdb_info").
 	Family("influxdb_uptime_seconds").
 	Family("http_api_requests_total",
-		L("handler", "platform"),
-		L("method", "GET"),
-		L("path", "/api/v2"),
-		L("status", "2XX"),
+		pr.L("handler", "platform"),
+		pr.L("method", "GET"),
+		pr.L("path", "/api/v2"),
+		pr.L("status", "2XX"),
 	)
 
 // UsageService filters prometheus metrics for those needed in the usage service.
 type UsageService struct {
-	Filter
+	pr.Filter
 }
 
 // NewUsageService filters the prometheus gatherer to only return metrics
 // about the usage statistics.
 func NewUsageService(g prometheus.Gatherer) *UsageService {
 	return &UsageService{
-		Filter: Filter{
+		Filter: pr.Filter{
 			Gatherer: g,
 			Matcher:  defaultMatcher,
 		},
